renames predefined_exps to predefined
